requirements.txt
groq
python-dotenv
pydub
requests
spacy
python-magic
regex